+++ .nvmrc
@@ -1,1 +1,1 @@
-v20.12.1
+v20.14.0
